Adjusted the handling of non standard GPU Memlets.
Once [PR#1976](https://github.com/spcl/dace/pull/1976) is merged in DaCe the code generator is able to handle more Memlets directly as Cuda `memcpy()` calls.
This PR modifies the GPU transformation of GT4Py in such a way that these Memlets are no longer transformed into Maps.

However, it can only be merged if the DaCe dependency was bumped to a version that includes PR#1976!