refactor[cartesian]: remove GT4PY_USE_HIP; break dependency cycle
This PR removes the GT4PY_USE_HIP configuration variable and uses the
CUPY_DEVICE_TYPE instead. This breaks the dependency cycle between
cartesian and storage. It removes the option to specify this variable as
an environment variable. According to Stefano auto-detection worked in
all cases so far. No need to keep this complicated.