feat[eve]: Preserve annex in custom visitors  (#1874)
Preserve annex attributes in the `NodeTranslator` for nodes returned
from custom `visit_` methods, not only from nodes which are recreated in
`generic_visit`. This is often desired and with this PR doesn't need to
be done manually in child classes of `NodeTranslator`. The `visit`
method can still decide if the node it returns should have a different
annex attribute as `_preserve_annex` does not overwrite in case the
attribute already exists.