feat[dace][next]: Updated Array Removal Transformations (#1915)
This PR updates how the redundant array removal transformations works.

First of all it updates the "self copy" transformation, i.e. where
information is copied from global memory into global memory, either
directly or indirectly. Because of ADR-18 we know that we can simply
remove such connections without doing an elaborated analysis.
There are two different transformations, that handles the case of a self
copy in the same state (there is also one that handles them across
multiple states):
- `SingleStateGlobalSelfCopyElimination`: This transformation was there
before. It essentially handles the pattern `(G) -> (T) -> (G)`, i.e. the
write is not direct but involves an intermediate. The transformation was
updated to handle more cases.
- `SingleStateGlobalDirectSelfCopyElimination`: This transformation was
newly added, it essentially handles the following pattern `(G) -> (G)`,
i.e. there is no temporary data involved.

In the long run these two transformation should be combined.

In addition the `CopyChainRemover` was updated.
Before it was required that the whole array, that should be removed,
henceforth called `A1`, was read, however, in some cases this could not
be checked symbolically.
To handle these cases, a new heuristic was implemented, which assumes a
certain behavior of the domain interference, i.e. that all arrays are as
small as possible, i.e. nothing that is not needed is computed.