fix[next][dace]: Fix array subset on the branches of if-expressions (#1912)
The issue was discovered by Philip (@philip-paul-mueller).

The lowering of if-expressions has to ensure that all branches (e.g.
if/then states) write the same subset to the result array. This was not
the case in baseline when one the branches contained a simple expression
such as `{dwdx, dwdy}`. The bug was that the full shape of the global
fields was copied into the result array, although the result array was
allocated on a sub-domain. This caused an out-of-bound write and a
segfault.