fix[next][dace]: Removed Workaround For DaCe Issue 1959 (#1921)
This [issue](https://github.com/spcl/dace/issues/1959) leaded to some
problem, especially in ICON4Py stencil `39_to_40`. However, through an
[update](https://github.com/GridTools/gt4py/pull/1915) in GT4Py this
leads to now to other problems (`ConstantPropagation` involving
AcccessNodes) in `19_to_20`. The underlying issue in DaCe was solved in
[PR#1980](https://github.com/spcl/dace/pull/1980) and this PR now
deactivate the workaround.

It also updates the version of DaCe.