feat[next][dace]: support as_fieldop with empty domain (#1929)
Zero-dimensional fields were supported in the dace backend as program
arguments, but not as temporary fields. This PR provides support for
zero-dimensional temporary fields, which are computed by `as_fieldop`
expressions with empty domain. This feature is enabled by PR #1853.

Note that the current representation in the SDFG as scalars does not
allow to return a zero-dimensional fields. However, the feature of
zero-dimensional fields as program output is not required.