feat[dace][next]: Added Aditional StateFusion Transformation. (#1927)
This PR adds the `GT4PyStateFusion` transformation which is a GT4Py
specific state fusion implementation.
This allows to handle some cases that are not supported by the one in
DaCe.
Because of the structure outlined in ADR-18 the transformation does not
check the subsets, but only checks the structure of the dataflow.