Fixed a bug edoardo has discovered.
It was caused because for some yet unknown reason, auto optimize runs twice.
And then the lcoal buffer transformation added a buffer again and did not modify the downstream.
This is s a genue bug that it solved by this PR, because now downstream data is corrected.
However, it is also made sure that the buffer is only added if the read is scalar.

Nevertheless, this whole file (`local_double_buffer.py`) must be rewritten as it is overcomplicated for a simple problem.