bug[cartesian]: fix setuptools for CUDA compilation (#1975)
In migration to versioningit (#1832 ), a `setuptools.cmdclass` is set
```python
[tool.setuptools.cmdclass]
# This is required for the `onbuild` versioningit hook
build_py = "versioningit.cmdclass.build_py"
sdist = "versioningit.cmdclass.sdist"
```
This overrides the `cmdclass` for CUDA that is passed to the
`pyext_builder` (instead of extending it).

Now, we set extend the `cmdclass` after the `Distribution` is created by
using internal hooks to first create the Distribution and then execute
it.

The code is not thread-safe!

---------

Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>