bug[next]: construct test meshes with correct allocator (#1974)
In #1963 @havogt missed to run the GPU tests and forgot that in tests we
actually never implemented proper device allocation for connectivities.

With this change the test mesh construction functions get a `allocator`
parameter. The `mesh_descriptor` fixture is now parametrized on
`exec_alloc_descriptor`. Tests that want to use the `mesh_descriptor`
with custom backends need to override the `exec_alloc_descriptor`.

---------

Co-authored-by: Edoardo Paone <edoardo.paone@cscs.ch>