Removed the `unit_strides_dim` argument from the optimizer.
The main reason is that `LoopBlocking` modifies the parameter names of the top Map.
Thus they are no longer identifiable through a matching.
However, the way how `LoopBlocking` creates the coarse name, the kind can still be used to identify it.
I also think that it was not rally usable anyway.