refactor[next]: Partial cleanup OffsetLiteralType (#1905)
When the type inference was introduced, the `itir.AxisLiteral` didn't
have a `kind` attribute. In order to still be able to assign a
`DimensionType` to them this property had to be deduced. After the
introduction of the `kind` attribute this is not required anymore and
this PR removes the deduction mechanism

---------

Co-authored-by: Till Ehrengruber <till.ehrengruber@cscs.ch>