fix[next]: Temporary extraction for fields defined on different domains (#1985)
Splits tuple values that are defined on different domains into separate
SetAts.

During temporary extraction we might extract a SetAt that contains a
tuple with elements defined on different domains (e.g. Vertex and Edge).
SetAts are required to have a single domain, therefore these SetAts are
split into one SetAt per domain at the end of temporary extraction.

Note: it seems many of the cases in icon4py are actually cases where the
domain is semantically equal, but the domain expression is different
(alla `1+1` != `2`). This can result in many unnecessary kernel calls.

---------

Co-authored-by: Hannes Vogt <vogt@hey.com>
Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>