Relaxed some restrictions.
Before the splitter required that for every producer there was at least one consumer, because this would indicate that we consume something that we do not use.
Which was before considered to be an error, at least I was thinking that.
However, we have updated MapFusion such that it can handle intermediates with multiple prducer.
This addition can create the fact that something is stored that is not read anywhere.
We now allow this, stuff that is nowhere read will then be removed by DaCe's dead datflow elimination.