Fixed a bug in the if mover regarding the removal of edges between nodes that are relocated and nodes that are not relocated.
The bug affected cases when an AccessNode that should be patched multiple times into the if block (different branches or same), has multiple connections to the MapEntry.
Then only the first connector is used, but the other ones are no longer needed.
Although the edges (inside the Map) are removed, the connector and the upstream dataflow is not.