fix[next]: Prevent extraction from lambdas in cse pass (#2013)
The common subexpression elimination pass already has a mechanism to
avoid extracting from `if_` calls, such that the evaluation of
expressions can be guarded in order to to avoid out-of-bounds accesses
or excessive pre-computation. The former protection can be circumvented
by placing things inside a lambda outside of the protecting if:
```
if cond then
  let f = λ() → ·guarded_it in
    if cond2 then f() else 0
else
  ·guarded_it
```
This PR prevents that by not extracting from lambdas unless they are a
let-form in which case it is clear that the function is unconditionally
evaluated.

---------

Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>