feat[next]: Precompile with static variants (#1968)
Adds the option to declare parameters as static (i.e. will be jitted)
and the option to (pre-)compile for scalar arguments.

Additionally, add `enable_jit` to the program. If `True`, variants for
the program will be compiled, if they don't exist yet. If `False` raises
an error if the program doesn't exist, i.e. wasn't pre-compiled.

The static parameters are static for all program variants
E.g. if param `a` and `b` are declared static (either explicitly or
implicitly by calling `compile`), then
- in the next call (with jit enabled) a new program will be compiled for
the values of `a` and `b`
- compile can only be called with static arguments for `a` and `b` (not
`a` or `b` only and not `c`)

---------

Co-authored-by: Enrique González Paredes <enriqueg@cscs.ch>
Co-authored-by: edopao <edoardo.paone@cscs.ch>