feat[dace][next]: Adjusted the handling of non standard GPU Memlets. (#1913)
Due to the update to DaCe in
[PR#1976](https://github.com/spcl/dace/pull/1976) the code generator is
now able to handle more Memlets directly as Cuda `memcpy()` calls. Thus
we have to change this function to reflect this case.
It even stands to reason that these functions can be removed altogether.

Note that the merge has become possible because of our switch to our
DaCe fork in [PR#2012](https://github.com/GridTools/gt4py/pull/2012).