feat[next]: Cache type synthesizer invocation (#2019)
The ITIR type inference is generally fast. In some narrow cases
encountered in icon4py where a lambda function is called many times the
runtime increases significantly since the return type is synthesized
many types even though it never changes (and can never change as all
lambda invocations need to have the same type by specification of the
IR). This PR restores linear runtime in the length of the expression by
caching the return type synthesis.

This was encountered in icon4py in
[`prepare_ffsl_flux_area_patches_list`](https://github.com/C2SM/icon4py/blob/82dd7cf9af5aa8506e20cb8bbafb6880f1b55ca2/model/atmosphere/advection/src/icon4py/model/atmosphere/advection/stencils/prepare_ffsl_flux_area_patches_list.py#L600)
after field operator fusion. The lambdas that are evaluated repeatedly
are created by `InlineCenterDerefLiftVars`.