Updated the `LoopBlocking` transformation.
The transformation now requieres that there are independent nodes before it applies.
Furthermore, that check was made more restrict, since it now filters out certain kind of nodes.
Mostly the ones related to `__out = 0.5` or so.
Furthermore, it now also requieres that the number of iteration is larger than the blocking.