feat[dace][next]: Modified How the GPU Block Size Is Set (#2020)
Before all GPU Maps had the same block size, although it was possible to
select different ones depending on the number of dimensions of a Map,
i.e. selecting `(128, 1, 1)` for Maps with just one parameter but `(32,
8, 1)` for Maps with 2 parameters.
This PR allows that the block size of Maps is adjusted downwards, if it
is known at code generation time, that the number of iterations is
smaller.

For an example assume that the block size is `(32, 8, 1)` and the Map,
for which we want to set the block size, only performs 4 iterations in
the `y` dimension, which is a common pattern with `concat_where` applied
to the vertical dimension.
Before we would have set the block size of that map to `(32, 8, 1)`,
which means that 50% of the threads would do nothing.
With this PR we will use a block size of `(32, 4, 1)`.