Fixed a problem reported by Edoardo related to long runtimes.
The match for parallel Map fusion is very general as it matches any two maps in the SDFG.
To reduce the search space this PR changes auto optimizer in such a way that first serial map fusion is performed on its own.
After that has ended, serial and parallel Map fusion are performed as before.

Note that the Map fusion transformation still performs the parallel matching, it is just always rejected, thus frequent rescans of the SDFG can be avoided, which is probably the reason for the speed up.
Furthermore, this PR also slightly modifies the `can_be_applied()` function of `MapFusion`, as it skips some steps.
It also removes the initialization of the edges from the `apply()` function, we assume that `apply()` is never called without a preceding call to `can_be_applied()`, which is a reasonable assumption.