Remove __INLINED deprecation note
While we all agree that `__INLINED` is not a nice name, there's
currently no alternative (and there won't be one anytime soon).
Deprecating a feature without an alternative and without a concrete
plan to remove it, is pointless and just clutters the log files.

We thus decided the remove the deprecation warning for now. Once there's
an alternative, we can put it back and push people to use the
alternative instead.