feat[next]: Infer as_fieldop type without domain (#1853)
- Extend the GTIR type inference to infer the type of `as_fieldop` calls
that don't have a domain argument.
- Avoids unnecessary promotion in the domain inference by using the now
available type information, e.g. `as_fieldop(deref)(vertex_field)` will
only get a `Vertex` domain also in cases when it is read from an
`as_fieldop` on a `Vertex, K` domain.
- The domain_inference of `as_fieldop` with `scan` was extended to keep
the vertical dimensions.
- `PruneBroadcast` was renamed to `RemoveBroadcast`, and the filename
was changed as well.

Motivation:
Previously, type inference was required to run after domain inference.
An `as_fieldop` call as emitted by the frontend had no domain, and
inferring the type of the domain was not possible, since there existed
no global ordering relation of dimensions, e.g. for an `as_fieldop`
operating on a `Vertex` and a `K` field it was unclear if the dimensions
of the output are `(Vertex, K)` or `(K, Vertex)`. This problem has many
negative consequences in other places:

- The domain inference was missing information on the dimensions of
intermediate / temporary fields and as a result might promote
unnecessary promote fields, e.g. an intermediate on `K` might be
promoted to `Vertex, K`.
- We could not run simplifications / transformations before inlining all
(named) functions as they require domain inference which did not support
them. As such the tree grows significantly before we simplifications,
unnecessarily increasing transformation / compile times.
- We had to run the collapse-tuple pass with partial type information,
as it is required to run before domain inference (in order to remove
field expressions that are never accessed).

The global ordering was already introduced in PR #1847.

---------

Co-authored-by: Till Ehrengruber <till.ehrengruber@cscs.ch>
Co-authored-by: Till Ehrengruber <t.ehrengruber@me.com>