fix[next][dace]: Enable mapping a scalar expression to a symbol on a nested SDFG (#2036)
The lowering from GTIR to SDFG was missing a feature: the possibility to
call a `Lambda` and map a scalar expression to a symbol used in an inner
domain expression. This corresponds to the mapping of a scalar data
container in the parent SDFG to a symbol on the nested SDFG, that is
used as range boundary or array shape inside the nested SDFG.

The lack of this feature represents a lowering issue in icon4py, for the
new combined stencils.