Unholy fix for Edoardo.
Then one can essentially only have one `sdfg.compile()` command in one function.
Thus we force the test suite to not use caching.
In addition we also add a function that compiles the SDFG in a way that ensures that it is generated anew.
Technically one would have to use this function anywhere, but we do it only partially here.