feat[next][dace]: Use dace cache for compiled programs (#2047)
The workflow step with the persistent `FileCache` store is translating
from `CompilableProgram` (ITIR.Program + CompileTimeArgs) to
`ProgramSource`, so this step is storing in cache only the result of the
SDFG transformations, not the compiled program binary.
With this PR, we rely on dace cache to avoid recompiling the SDFG every
time the application is launched.

---------

Co-authored-by: Philip Mueller <philip.mueller@cscs.ch>
Co-authored-by: Philip Müller <147368808+philip-paul-mueller@users.noreply.github.com>