feat[next][dace]: Add config option for async sdfg call (#2035)
This PR introduces a config flag `async_sdfg_call` (default `False`) on
the translation stage to enable asynchronous SDFG execution on GPU.

By default, the call to a dace SDFG is synchronous: it does not return
until the execution has completed. This is implemented in the dace
generated cuda code by `cudaStreamSynchronize()`. The newly added
`async_sdfg_call` flag allows to avoid these synchronization calls, by
executing all kernels on the default cuda stream and rely on queue
serialization. In this way, the SDFG call will immediately return and
kernel execution on GPU can overlap with python execution on CPU. This
approach is already used by the GTFN backend and allows to hide the
latency of the python host code.

In the translation stage of the dace backend, when
`async_sdfg_call=True` and the SDFG contains some GPU maps, the SDFG is
modified as follows:
- The init code is extended with a call that sets all cuda streams to
the default cuda stream.
- The states of the top-level SDFG are modified with `nosync = True`
that will skip generating `cudaStreamSynchronize()` in the target code.