feat[next][dace]: Add SDFG transformation for map splitting (#1992)
This PR introduces 2 new transformations: `gt_horizontal_map_fusion` and
`gt_vertical_map_fusion`

The objective of these transformations is to identify maps with
partially overlapping range and extract the range intersection as a
separate map, in order to enable map fusion. The horizontal and vertical
transformations differ in the pattern they look for:
- horizontal looks for any two maps with a common ancestor that can be
fused by parallel fusion
- vertical looks for serial maps (i.e. two maps connected by some data
container in between) in order to promote serial map fusion

However, several pieces are common, in particular the utilities for
splitting the map range and copying a map scope.

---------

Co-authored-by: Ioannis Magkanaris <ioannis.magkanaris@cscs.ch>
Co-authored-by: Philip Mueller <philip.mueller@cscs.ch>
Co-authored-by: Ioannis Magkanaris <iomagkanaris@gmail.com>
Co-authored-by: Philip Müller <147368808+philip-paul-mueller@users.noreply.github.com>