fix[next][dace]: Updated `SerialMapPromoter` (#2057)
This PR modifies the `SerialMapPromoter` in such a way that it performs
the fusing of the two Maps immediately after the promotion.
This is needed because technically such an SDFG is invalid as memory
locations are now written multiple time (it was not critical, because
they always write the same value).
By performing the fuse inside the `apply()` function, no invalid SDFG is
visible to the outside.

Furthermore, this PR removes the `BasePromoter` that was not used
anywhere else.
As an additional restriction it is now only possible to promote the
first Map, and no longer the second Map, however, this feature was
probably unneeded anyway (in GPU transformations it might be needed).