fix[next][dace]: Add method to create custom dace backend (#2059)
This PR was developed while debugging why the gt4py persistent cache was
not working in the icon4py granule tests with the dace backend. It was
found that the granule test was setting up the dace backend with custom
config in a wrong way, basically disabling the translation cache.
Therefore, this PR introduces a function `make_dace_backend()` to create
a dace backend with custom config for SDFG lowering and auto-optimize.

Additional changes contained in this PR, still related to execution of
the granule with dace backend:
- Some cleanup in the area of SDFG serialization to JSON and
compilation.
- Fix handling of SDFG scalar arguments for field shape/stride/domain.